Move MaterialSlot.java to jp.mincra.mincramagics, as it is used by multiple packages.
